update enc_cipher_ok for growing key heaps
To accommodate key generation, we need to show that encrypted_cipher_ok
is preserved by the addition of another honest key. This requires that
the key is *actually* new, i.e. that it could not have been used to
sign/encrypt a cipher previously blessed by encrypted_cipher_ok.

We therefore include the global key heap as a parameter to
encrypted_cipher_ok and stipulate in its constructors that:
1) the signing/encryption keys must be in the global key heap
2) any (private) keys in the message to be packaged must be in the
global key heap.

Then to show preservation, we need only know that the new key was not
previously in the global key heap (a fact which follows directly from
the key generation steps in the operational semantics).

I've also included some automation which may or may not have application
in other proofs where key_perm manipulation is required.