use type codes in protocols
I don't love the organization of this currently, but all the stuff I
was trying with typeclasses didn't end up much nicer since ultimately we
want to segregate base types from things like messages and crypto.

Currently one can only create universes with base types. In our setting
this should be fine since we want the real/ideal universes to be
comparable and the base types are precisely the types the impl/spec
languages share.